requirements.txt
azure-ai-evaluation
azure-search-documents
azure-identity
python-dotenv
